Fix a deadlock situation when the GC wants the stop the world
When the GC tries to run, it stops all running threads. Whenever
there was contention over a certain mutex used in a LockGuard, it
could happen that a thread was trying to get a lock when the
LockGuard was initialized. In that case the thread wasn't marked
as GC independent, so the thread that wanted the world to stop
for the GC would never see that all threads were waiting. At that
point all threads would be waiting forever.

Fixes #1110