Imported matrix.rb from MRI r33295.
Since many of the failing Matrix specs are guarded with ruby_bug, we either
must fail them or fix them even though we are targetting 1.9.2. The fixes
would not seem to cause problems for Matrix users, but if they do, we can
revisit whether to import from 1.9.2 and fail the specs or update the specs
to use ruby_version_is rather than ruby_bug. There is a fine line between
a version difference and a bug. I tend to favor only using ruby_bug when
the value computed is clearly in error or there is a hang or segfault.