Add ability for the JIT to specialize toplevel methods
When the JIT compiles a method, it can compile it for a staticly known
self class. This allows things like self sends and ivar access to be
much faster. For now, a special specialized_executor is needed to pick
from the avaliable JITd methods the proper one for the receiver.

Future inline cache work will allow us to cache the specialized version
in the call site, eliminating the need to hit the specialized_executor
on every send.