Fix incorrect running of finalizers bug
Baker::check_finalize checks if an object is forwarded to know if it has
still alive or not therefore whether or not run it's finalizer.
Therefore we can not check finalizers more than once because the 2nd
time, we'll see objects that we have fixed up (and therefore are not
forwarded) and finalizer them.