Use Object if no class is given to rb_data_object_alloc
MRI allows for creating a very strange object that has
no klass pointer set. This means this object can't for example
be inspected, since it will given the following result:

NotImplementedError: method `class' called on terminated object (0xXXXXXXXX)

In Rubinius we just use Object, so it at least works. It is
definitely recommended to use a proper actual class to prevent
potential issues on MRI.

Fixes #1187