Audited usage of char* for Symbols
Since Symbols can contain \0 characters in 1.9, we need to make sure
this is handled properly. There is now an explicit debug_str method
used for when outputting for example symbol names. This returns
and escaped std::string for the symbol name.

For other explicit usages of the Symbol name, there is cpp_str to return
a std::string without the escaped non printable characters. Most cases
use the debugging version though.