patch String#squeeze for 19
String#squeeze should throw an ArgumentError when given a bad range
(e.g., 'e-b').

Since String#squeeze and String#squeeze! should both throw an
ArgumentError when the range argument is backward, I moved the error
into the String#squeeze! method and removed it from the String#squeeze
method (which is defined in terms of the banged method).