Removed C++ tests for primitives tested in RubySpec.
The C++ tests are most useful for actual VM behavior that is opaque to Ruby.
In the case of primitives, they are (or should be) soundly tested by RubySpec
and breakage should be immediately visible. Further, the primitives are
implementing Ruby behavior, which may change between versions, so the overhead
to managing that is very high in the C++ tests and very easy in RubySpec.