Fix Fiber/VariableScope GC bug
The GC system assumes that any Callframe* in the system is walked every
GC. A mature Fiber would thus violate this assumption, which in term
broke a VariableScope that pointed into the CallFrame* chain associated
with the mature Fiber because the VS's locals_ were not seen by the GC
(since the VS was mature too). Because VS::locals_ can reference a
StackVariable's storage directly, no write barrier is run.

Thusly, we must touch every Fiber every GC so that the CallFrame* chain
assumption is upheld.