Added CI tags for encoding related specs.
The parser was fixed to use a better function API for encoding compatible
functions on Rubies that don't support encodings. This change uncovered
some specs that were erroneously passing (ie, passing because the wrong
compatibility code was being used and the correct compatibility code has
errors that need to be fixed still).