Make sure to leave / enter C-API around usage of rubinius::Helpers methods
Using a rubinius::Helpers methods calls back into Rubinius and
does not involve C-API methods. This means we should unlock
the C-API lock because those helpers might for example call an Autoload
which ends up compiling a new file and wants to grab the C-API lock
for the parser extension.

Fixes #1402