Allow using a different linker in Daedalus
Not on all platforms (e.g. Solaris) you can link with gcc
and expect all c++ features to work. If you don't use g++ on
Solaris, c++ exceptions won't work and this Rubinius crashes.

This changes allows Daedalus for using a different linker and
sets it up to use g++ by default so we link the Rubinius binary
with g++.