On Solaris getcwd allocates a buffer of the given size
If you use getcwd on Solaris with a NULL pointer as the
first argument, it will only allocate a buffer the size
of the second argument when creating it. By giving a
value larger than 0 as the argument, we make sure that
enough space can be allocated for actually returning a
String in this test.