Use the correct type for htons according to the man page
This actually breaks otherwise on Solaris, since there is
no type u_int16_t on that platform. The man page for htons
also names uint16_t as the type for the argument and return
value so we should adhere to that type.

The previous commit makes sure this type is actually discovered
on OS X, since on that platform apparently including sys/types.h
doesn't make this type available.