Validate a MCE in all check_cache functions
IC::execute_backend_ and IC::cache_ are mutated seperately but their
values are intertwinned. Without this fix, a cache_ that points to
Array#inspect could be used for a Fixnum because check_cache_fixnum
didn't validate that the mce was actually for a Fixnum method. This
happens because multiple threads mutate the IC at the same time.

This obviously causes crashes.

This fix at least makes sure that the shortcut check_cache_* functions
are using a cache thats allocated with the proper check_cache_*
function.