#send/#__send__ must do all coercion in the primitive
Coercing Ruby-side and then re-calling #send/#__send__ would
produce incorrect results when sending messages that are sensitive
to the call stack like send("__callee__") and the regex globals
following send("gsub").

The previous retry code also failed to pass the block parameter. It
was essentially unreachable anyway, but for clarity I think this is
better.