Add deprecated headers like MRI does
This shows a warning when using these headers, but there are
packages such as bcrypt-ruby that still use these deprecated
headers. The reason that bcrypt-ruby usually doesn't fail is
that it includes util.h which is a system header on most systems.

If a system didn't have a util.h present it would fail to install
like in issue #1445.

Fixes #1445