Add specs for two edge cases regarding Bignum bitwise operations
The Bignum representation of these numbers on Rubinius resulted
in them being represented with all 1's. Of those are subsequently
inverted for 2's complement, it turns into all 0's. We need to
make sure not to clamp the values in those cases because that
will result in different answers.

These two specs verify this for two cases that result in these
kind of values for both 32 and 64 bit platforms.