Fixed defining encodings.
We switch from a map of {symbol -> index} to {symbol -> [alias_name, index]}.
This is necessary to correctly list the names of aliased encodings.  There is
no way to create an alias from Ruby and the aliased encoding actually returns
the original encoding when referenced, but the aliased name is shown in the
Encoding.name_list results. For example:

  Encoding.find("binary").name # => "ASCII-8BIT"

but "BINARY" must be listed in Encoding.name_list results. Further
complicating this is that encodings must be case-insensitive searchable.