Guard for a very nasty memory corruption bug
If someone already used all bytes in a CharArray that is
backing a String object, setting a trailing \0 to this
CharArray will overwrite memory outside the CharArray.

This leads to other objects being corrupted and leads
to crashes of the VM.

This issue is exposed because of how String#splice! works.
If you happen to use it in a such a way that you fill up
the exact number of bytes in a CharArray and subsequently
that CharArray is used in a method that uses c_str() under
the hood, it corrupts memory. An example of this behavior
can be seen in String#succ:

s = "zzzzzzz".succ
s => "aaaaaaaa"
s.data => #<Rubinius::CharArray:0xf88 8 bytes>
s.to_sym => Rubinius::ObjectBoundsExceededError: can't safely create a null terminated string

Before this guard, this code would corrupt the memory
by writing a \0 outside the CharArray. This would lead
to unexpected memory corruption and for example GC crashes
which is where I first saw this.

Happy holidays everyone!