Removed CharArray. Use ByteArray in String.
Initially, the thought was that a CharArray could encapsulate the idea of a
vector of bytes and the interpretation of those bytes relative to a particular
encoding scheme. However, in practice, the interpretation of those bytes is
really encapsulated in String, which composes a ByteArray and an Encoding.
Pushing the logic down into CharArray required delegating almost everything
from String, which is a good indicator for a poor abstraction.

One example in particular illustrates this: a ByteArray (and CharArray)
contain a boundary-aligned number of bytes, the boundary being a machine word.
The size of a ByteArray (CharArray) is always >= to the number of bytes needed
for a String's data. Encoding operations need to operate on the precise number
of bytes in the String's data because those extra bytes that pad to a boundary
in a ByteArray would be misinterpreted in some Encodings.

Essentially, the more Encoding-aware CharArray became, the more it was just a
String under String. So we removed it.