made alloca.h includes only for solaris
In September alloca.h was added to build on solaris systems. The problem
there was, that alloca.h was not included through stdlib.h.

FreeBSD has no alloca.h which can be included, because everything is in
stdlib.h and therefore rubinius could not be built.

On Linux, it's no problem, as alloca.h gets included through stdlib.h.

So the include for alloca.h is now only for solaris systems.