Safeguard casting founded encodings to Encoding
The problem here is that it might happen that we get back
a Qnil if an encoding can not be found. Using as<> causes
this to raise a TypeError, but raising that error doesn't
properly unlock the CAPI lock. This would cause a deadlock
if after this happened another call into the CAPI was made.

This was happening when running pg gem specs.