Use version-specific helpers instead of version conditionals.
Does this approach duplicate some code? Yes. But it does so at a minimum
logical level. Duplicating all of #eval is too high. Duplicating some logic
inside one "coercion"-type method is about right, the concept is better
contained (encapsulated).

The biggest problem with conditionals is that we must support multiple Ruby
language versions, not just 1.8 and 1.9. Using version-specific helper methods
keeps the code consistent and clear. Think of them as a form of polymorphism.