Replaced RTEST in the vm with CBOOL.
RTEST is a macro defined in MRI that converts a Ruby object to a C boolean
value. The name itself is not that great from the perspective of revealing
intention, but most importantly we have to have vm code and C-API code
co-existing and we can't do that by overloading the meaning of RTEST (because
Qtrue, Qfalse, Qnil are different values in the vm and in the C-API).

There is some precedent around for the name CBOOL but I don't care what the
name is, as long as it is not RTEST. We should probably use true_p() and
false_p() uniformly instead.