Cleanup arity checks for blocks as methods
This maybe somewhat more verbose than the older implementation
but this is a lot easier to understand what is happening. It took
me quite some time to understand the old if statement so therefore
I rewrote it.

This also fixes the case where a block has default arguments in 1.9
mode.

Fixes #1502