Use atomic operations on pending_threads and spin when stopping the world
This greatly improves performance of the gc_dependent / gc_independent
barrier. This barrier is often crossed for potentially blocking and
GC indepedenent functions such as FFI and IO calls.

This makes these calls a lot cheaper because in the common case of not
needing to GC they only have to automatically increase and decrease
the pending_threads_ counter instead of doing a full lock / unlock.