Tag a stupid platform-dependent spec.
This is the MRI behavior where values larger than a Fixnum that still fit in a
machine integer are used in C code. This is basically opaque to Ruby code and
is generally a terrible idea. In many cases, Rubinius requires a Fixnum except
where interfacing with libc functions whose valid range is a machine integer.

I need to add machine integer helpers to MSpec for cases where this is
reasonable. The other cases are typically marked not compliant on Rubinius.