Register Inliners properly as a CodeResource so it can be cleaned up
This means we also need to mark Inliners which now happens in the
mark_inliners function used by various mark functions for different
Executable's.

This also showed that NativeFunction wasn't properly marking it's
inliners which probably never led to any real problems, but with
the change to track the Inliners properly, this did expose this
lingering issue.

Also removed some redundant checks since we just asserted in the
line above that inliners_ isn't null.