Fix two 1.9 YAML specs by calling #to_time on DateTime objects
When trying to replicate the functionality that the specs were testing
on MRI, the result of calling YAML.load was a Time object, not a
DateTime object like in Rbx. The DateTime class doesn't implement #usec,
so naturally the specs were failing. Making YAML.load return a Time
object fixes the specs without breaking any of the others.