Only install specialized accessors for classes, not modules
This can cause problems because accessors in a Module can't
assume that the receiver class is always the same here. If
you use the module with accessors in different classes it
would specialize the accessor and would run the wrong code
if for example an immediate would be used with the module
instead of a reference the first time.

Fixes #1567