Fixed Module#name.
OMG Module#name. We have moved a lot of critical functionality into
Rubinius::Type to avoid issues with people undef'ing methods on core
library classes. We need to create a proper Mirrors implementation.
So some of this commit removes helper methods and under, under protective
methods for Module#name. Further, we need the name at various points in
the VM/C-API, so the 1.8/1.9 logic for "" vs nil for an anonymous module's
name is now in the primitive.