Don't raise after error in getnameinfo()
When the computer isn't connected to any network, some specs fail only on
Rubinius. They don't on MRI.

The cause was that Rubinius's getnameinfo() is slightly different from MRI's one
in the corner case of error handling.

So, Rubinius's getnameinfo() is changed to be aligned with MRI's.

One of the failing specs is this:

  Socket#getaddrinfo accepts empty addresses for IPv4 passive sockets ERROR
  SocketError: Temporary failure in name resolution