Remove a VM from the threads list when discarding it
Previously there were cases where we weren't properly removing the
thread from the threads list. This happened for example for the
SignalHandler thread and QueryAgent thread.

By always removing it from the threads list from VM::discard, we make
sure it's always remove and does not cause a crash when the GC tries to
scan a thread that was not running anymore.