Split the actual execution of the coerction into separate methods
This makes them nicer to the JIT, which can now better inline the first
step where just the type is checked. Since the common case is that we
don't have to coerce, this makes it cheaper for those cases. The
complete methods often use exceptions which makes it harder to inline.

This improves performance, for example for Array#dup:

Before
=== bin/rbx ===
         dup strings  3492520.3 (±3.8%) i/s -   17446494 in   5.003134s (cycle=123734)
         dup numbers  3471273.3 (±4.4%) i/s -   17330250 in   5.003426s (cycle=165050)

After
=== bin/rbx ===
         dup strings  3664231.1 (±1.6%) i/s -   18353628 in   5.010230s (cycle=124011)
         dup numbers  3656536.5 (±3.9%) i/s -   18368316 in   5.032562s (cycle=173286)

We probably still want to look into making the coercion checks cheaper,
since they have quite an impact still.