Apply the VM-to-State transition to the JIT code
Since the following commit, many C++ functions are changed to take a pointer to
a State, not a pointer to a VM. But the LLVM code hasn't catch up the change.

  35a49f3 Introduce State class as STATE

This is very confusing. Just update it.

The reason Rubinius doesn't currently break in this situation is that the LLVM
IR type system is completely independent of the actual C++ type system.