Fix Hash order dependency in code generation
On Ruby 1.8 implementation such as Rubinius and the latest patchlevel
1.8.7 Hash ordering is much more random. This means that hashes are
ordered differently each time and that thus the code generation here
ended up generating the code in different order.

The result of this was that when you use one of these Rubies to build
Rubinius, each time the whole VM was being rebuilt instead of only the
necessary parts since the generated source files changed each time.