Correctly initialize TERMINAL_WIDTH
Currently, Rubinius::TERMINAL_WIDTH is always set to 80, not to the actual
number of columns for the current terminal.

This commit fixes this bug, which is introduced by this commit:
  87c1849 Guard querying tput more closely

Rubinius::TERMINAL_WIDTH is set by reading the output of the "tput" command.
However, tput always prints "80", because it can't access to the current
terminal by any means. It's because tput's stdout nor stderr is connected to
the current terminal's file descriptor. Both are redirected by rubinius.

This commit allows tput to access to the current terminal again via stderr and
can be confirmed by the following strace invocations:

$ strace -f -e trace=ioctl ./bin/rbx -e ''

Before this commit:
[pid 11709] ioctl(1, SNDCTL_TMR_TIMEBASE or TCGETS, 0x7fff39d857c8) = -1 EINVAL (Invalid argument)
[pid 11709] ioctl(2, SNDCTL_TMR_TIMEBASE or TCGETS, 0x7fff39d857c8) = -1 EINVAL (Invalid argument)
[pid 11709] ioctl(2, SNDCTL_TMR_TIMEBASE or TCGETS, 0x7fff39d85778) = -1 EINVAL (Invalid argument)

After this commit:
[pid 11293] ioctl(1, SNDCTL_TMR_TIMEBASE or TCGETS, 0x7ffff444c138) = -1 EINVAL (Invalid argument)
[pid 11293] ioctl(2, SNDCTL_TMR_TIMEBASE or TCGETS, {B38400 opost isig icanon echo ...}) = 0
[pid 11293] ioctl(2, SNDCTL_TMR_TIMEBASE or TCGETS, {B38400 opost isig icanon echo ...}) = 0
[pid 11293] ioctl(2, TIOCGWINSZ, {ws_row=43, ws_col=227, ws_xpixel=0, ws_ypixel=0}) = 0