Support Regexp::{FIXEDENCODING,NOENCODING}
Since the numeric value of these constants is exposed via
the constant itself and via Marshal.dump, I thought it best
to use the same values as MRI. Even though client code
_should_ always use the symbolic constant, there's no
guarantee it does.

Doing so required changing the values used for the legacy
kcode options. Since they were specific to Rubinius, this
should have minimal consequences. We just need to be sure
that they are kept in sync between the kernel, C++, and
melbourne.

Also, remove the 1.8 Regexp#kcode method from 1.9.