fix the order of include paths
When oniguruma was installed on a system and located in the include path
of llvm, then the compile could brake. This happend on freebsd 9.0 with
oniguruma and llvm installed from ports.

This fix moves the inclusion of the paths after all libs are processed
so that these include paths are first.