Use union instead of struct for cmsghdr
Use of struct with cmsghdr and pad is a GNU extension. When built with
Clang, warnings are issued as such. So take the union approach instead.

To fix this, I used an example at:
  http://www.rootr.net/man/man/CMSG_SPACE

This is a second try to fix clang warnings in io.cpp. The reverted first try
was:
  a9648cfc771fd189471825c42dcd28c499fded47 Ignore all GNU-extension warnings

As an alternative approach to it, I fixed the code itself.