Use separate spinlocks for memory and inflated header allocation
We also remove the lock around setting object id's by using an atomic
increment. This does introduce the possibility that object id's are
skipped because the same object can get an object id concurrently. This
however is a very minimal case since it means that you are creating new
object id's on constantly modifying global state.