Setup rb_usascii_str_new_cstr as an alias for rb_usascii_str_new2
We already had rb_usascii_str_new_cstr, but MRI defines this as an alias
for rb_usascii_str_new2. We just mimic this and rename the old method
and create a define for rb_usascii_str_new_cstr just like MRI.

Fixes a build issue with the new Psych version.