Fix memory leak and race condition in agent
There were two problems here. One is that we never joined on the thread
started for the agent which causes a memory leak in Linux. Although the
leak is small, it leaks a lot more in virtual memory size. This causes
virtual memory sizes for the rbx process of up to 3GB in spec runs.
Since this is pretty close to the maximum on 32 bit systems, this can
cause memory errors on those platforms.

Ensuring that we join the thread makes sure we don't leak this virtual
memory and our process size stays around the same values as running
without the agent enabled.

The other problem is a race condition in setting up the local loopback.
Here we start the agent thread that then goes waiting for events. The
problem is that the loopback sockets aren't setup until after starting
the thread. This causes the race possibility where the agent is already
calling select() for the sockets, without having the loopback sockets in
the fd_set. To work around this issue, we trigger a wakeup for the agent
after setting up the loopback sockets so that the loopback sockets are
also monitored for activity.