Fix Autoload handling by properly reusing the current scope
In the Autoload parts in the VM we now do the lookup again so it's much
easier to do it with the correct current scope and we follow the same
lookup mechanisms as a normal constant lookup.

For this we modify the Helpers for const_get so it can filter a result
so the autoload is filtered and the real class or a class with the same
name higher in the hierarchy is returned.