Add spec for Struct with attribute name containing hyphens.
Currently Rubinius allows to create Struct instances with
attribute names containing hyphens (e.g., 'next-state' or
'age-range'), but calling #to_a or #instance_variables
on such struct objects raises a runtime error.

The following example works both on MRI and on previous
releases of Rubinius:

    klass = Struct.new(:'current-state', :symbol)
    tuple = klass.new(0, :>>)

But currently fails with a NoMethodError for `-'.